requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
google-genai
pillow-heif
ultralytics
python-multipart
scikit-learn
seaborn